automatically backed up by learn

+++ app/models/haunted_house.rb
@@ -5,7 +5,7 @@ class HauntedHouse < ActiveRecord::Base
       h.string :name
       h.string :location
       h.string :theme
-      h.integer :price
+      h.float :price
       h.boolean :family_friendly
       h.datetime :opening_date
       h.datetime :closing_date
